refactor(tabs): rename teamsfx to teams dev
