+++ tabs/src/components/teamsdev/bot/AdaptiveCard.jsx
