+++ tabs/src/components/teamsdev/bot/WelcomeCard.jsx
