+++ tabs/src/components/teamsdev/tab/StaticTab.jsx
